pyproject.toml
[project]
name = "poc-sisu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.0",
    "langchain-openai>=1.1.6",
    "langchain-core",
    "langgraph>=1.0.5",
    "pinecone>=6.0.0",
    "tiktoken>=0.12.0",
    "python-dotenv",
    "langfuse",
]
